requirements.txt
ipython==3.2.1
cocos2d==0.6.3
